refactor app layout
