+++ lib/src/constants.dart
@@ -1,0 +1,2 @@
+const String kSmallLogoImage = 'logos/pesto/logo_small.png';
+const String kGalleryAssetsPackage = 'flutter_gallery_assets';
